scst: Remove a superfluous cast



git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@4480 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/src/scst_lib.c
@@ -6430,7 +6430,7 @@ static int get_cdb_info_len_1(struct scst_cmd *cmd,
 	cmd->op_flags |= SCST_LBA_NOT_VALID;
 	cmd->lba = 0;
 
-	cmd->bufflen = (u32)cmd->cdb[sdbops->info_len_off];
+	cmd->bufflen = cmd->cdb[sdbops->info_len_off];
 	cmd->data_len = cmd->bufflen;
 	return 0;
 }
